+++ packages/dodam-design-system/src/navigations/utils/path-from-stack.ts
