+++ packages/dodam-design-system/src/navigations/utils/route-score.ts
