+++ packages/dodam-design-system/src/navigations/utils/segments.ts
