+++ packages/dodam-design-system/src/navigations/utils/sort-routes-by-priority.ts
@@ -1,5 +1,5 @@
 import { scoreRoute } from "./route-score";
-import { RouteConfig } from "./types";
+import { RouteConfig } from "../core/types";
 
 export const sortRoutesByPriority = (routes: RouteConfig[]): RouteConfig[] => {
   return [...routes].sort((a, b) => scoreRoute(b.path) - scoreRoute(a.path));
