+++ packages/dodam-design-system/src/navigations/utils/stack-from-path.ts
